add - doc - Added support for sorting 128-bit integers
---

You can now sort 128-bit integers using the ArrayTools functions, too!

---

Type: add
Breaking: False
Doc Required: True
Part: 1/1

+++ public/Nitrocid/Drivers/Sorting/BaseSortingDriver.cs
@@ -17,6 +17,8 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 //
 
+using System;
+
 namespace KS.Drivers.Sorting
 {
     /// <summary>
@@ -157,7 +159,7 @@ namespace KS.Drivers.Sorting
         }
 
         /// <inheritdoc/>
-        public long[] SortNumbersInt128(long[] unsorted)
+        public Int128[] SortNumbersInt128(Int128[] unsorted)
         {
             // Get the number of iterations
             int iteration = unsorted.Length;
